feat(#783): ICA: tsc-checkpoint monolithic — extract adapter/watcher/format modules
