--- .pi/extensions/tsc-checkpoint/test/tsc-checkpoint.test.mts
@@ -1,1728 +1,0 @@
-/**
- * Tests for tsc-checkpoint
- *
- * Phases 1: diagnosticToTscDiagnostic
- * Phase 2: runTscCheckpoint one-shot (ts.createProgram) integration
- * Phase 3: TypeScriptWatchAdapter refactoring (delegation)
- * Phase 4: DiagnosticsWatcher — Lifecycle
- * Phase 6-8: DiagnosticsWatcher incremental cache, file-path resolution, trends
- * Phase 9: Pipeline contract — runTscCheckpoint signature
- * Phase 10-15: Extension entry point, lifecycle, resource leaks, trust gate, modes, parseArgs
- *
- * Phases 4-8 use MockAdapter to keep tests fast and deterministic.
- * Phase 2 uses real ts.createProgram with temp fixtures.
- *
- * Run with:
- *   node --experimental-strip-types --test .pi/extensions/tsc-checkpoint/test/tsc-checkpoint.test.mts
- */
-
-import assert from "node:assert";
-import { describe, it, beforeEach } from "node:test";
-import { resolve } from "node:path";
-
-import {
-	DiagnosticsWatcher,
-	resolveDiagnosticFilePath,
-	formatDiagnostics,
-	formatDiagnosticsJson,
-	formatTrend,
-	runTscCheckpoint,
-	diagnosticToTscDiagnostic,
-} from "../index.ts";
-
-import type {
-	TscDiagnostic,
-	TscWatchAdapter,
-	TscCheckpointResult,
-	DiagnosticTrend,
-} from "../index.ts";
-
-import { mkdtempSync, writeFileSync, rmSync, mkdirSync } from "node:fs";
-import { join } from "node:path";
-import { tmpdir } from "node:os";
-import ts from "typescript";
-
-// ═══════════════════════════════════════════════════════════════════════
-// Mock Adapter for Testing
-// ═══════════════════════════════════════════════════════════════════════
-
-class MockAdapter implements TscWatchAdapter {
-	startCalls = 0;
-	stopCalls = 0;
-	lastStartPath = "";
-	private _isRunning = false;
-	private _diagnostics: TscDiagnostic[] = [];
-	private _listeners: Array<(diagnostics: TscDiagnostic[]) => void> = [];
-	private _shouldFailStart = false;
-
-	setShouldFailStart(fail: boolean): void {
-		this._shouldFailStart = fail;
-	}
-
-	start(tsconfigPath: string): boolean {
-		this.startCalls++;
-		this.lastStartPath = tsconfigPath;
-		if (this._isRunning) return false;
-		if (this._shouldFailStart) {
-			throw new Error(`tsconfig not found: ${tsconfigPath}`);
-		}
-		this._isRunning = true;
-		return true;
-	}
-
-	stop(): void {
-		this.stopCalls++;
-		this._isRunning = false;
-	}
-
-	isRunning(): boolean {
-		return this._isRunning;
-	}
-
-	getDiagnostics(): TscDiagnostic[] {
-		return this._diagnostics;
-	}
-
-	onDiagnosticsChange(callback: (diagnostics: TscDiagnostic[]) => void): void {
-		this._listeners.push(callback);
-	}
-
-	/** Test helper: simulate a diagnostic event from the watch process */
-	emitDiagnostics(diagnostics: TscDiagnostic[]): void {
-		this._diagnostics = diagnostics;
-		for (const listener of this._listeners) {
-			listener(diagnostics);
-		}
-	}
-}
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 1: diagnosticToTscDiagnostic — pure mapping ts.Diagnostic → TscDiagnostic
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("diagnosticToTscDiagnostic", () => {
-	const configDir = "/home/user/project";
-
-	function mockSourceFile(fileName: string): ts.SourceFile {
-		return {
-			fileName,
-			getLineAndCharacterOfPosition(_pos: number) {
-				return { line: 2, character: 5 };
-			},
-		} as unknown as ts.SourceFile;
-	}
-
-	function mockDiagnostic(overrides: {
-		file: ts.SourceFile;
-		start?: number;
-		messageText?: string | ts.DiagnosticMessageChain;
-		code?: number;
-		category?: ts.DiagnosticCategory;
-	}): ts.Diagnostic {
-		return {
-			start: 100,
-			messageText: "Type 'string' is not assignable to type 'number'",
-			code: 2322,
-			category: ts.DiagnosticCategory.Error,
-			...overrides,
-		} as unknown as ts.Diagnostic;
-	}
-
-	it("maps error diagnostic with file to correct TscDiagnostic fields", () => {
-		const file = mockSourceFile("src/app.ts");
-		const diagnostic = mockDiagnostic({ file });
-
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-
-		assert.ok(result, "should return a TscDiagnostic");
-		assert.strictEqual(result!.file, "src/app.ts");
-		assert.strictEqual(result!.line, 3); // line + 1
-		assert.strictEqual(result!.column, 6); // character + 1
-		assert.strictEqual(result!.severity, "Error");
-		assert.strictEqual(result!.message, "Type 'string' is not assignable to type 'number'");
-		assert.strictEqual(result!.code, "TS2322");
-		assert.strictEqual(result!.filePath, "/home/user/project/src/app.ts");
-	});
-
-	it("maps diagnostic with non-zero offset → line/column derived correctly", () => {
-		const file = {
-			fileName: "src/deep.ts",
-			getLineAndCharacterOfPosition(pos: number) {
-				// pos 150 → line 3, character 30
-				return { line: 3, character: 30 };
-			},
-		} as unknown as ts.SourceFile;
-
-		const diagnostic = mockDiagnostic({ file, start: 150 });
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-
-		assert.ok(result);
-		assert.strictEqual(result!.line, 4); // 3 + 1
-		assert.strictEqual(result!.column, 31); // 30 + 1
-	});
-
-	it("maps diagnostic with nested messageText → flattened single string", () => {
-		const file = mockSourceFile("src/app.ts");
-		const nestedMessage: ts.DiagnosticMessageChain = {
-			messageText: "Type 'string' is not assignable",
-			category: ts.DiagnosticCategory.Error,
-			code: 2322,
-			next: [
-				{
-					messageText: "Did you mean 'number'?",
-					category: ts.DiagnosticCategory.Error,
-					code: 2322,
-				},
-			],
-		};
-
-		const diagnostic = mockDiagnostic({ file, messageText: nestedMessage });
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-
-		assert.ok(result);
-		// Flattened message should include both parts
-		assert.ok(result!.message.includes("Type 'string' is not assignable"));
-		assert.ok(result!.message.includes("Did you mean 'number'?"));
-	});
-
-	it("diagnostic without file → returns undefined", () => {
-		// A diagnostic without a file (global error like duplicate identifier across files)
-		const diagnostic = {
-			start: 0,
-			messageText: "Global error",
-			code: 2300,
-			category: ts.DiagnosticCategory.Error,
-			// No file property
-		} as unknown as ts.Diagnostic;
-
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-		assert.strictEqual(result, undefined);
-	});
-
-	it("diagnostic with zero start → line=1, column=1", () => {
-		const file = {
-			fileName: "src/zero.ts",
-			getLineAndCharacterOfPosition(_pos: number) {
-				return { line: 0, character: 0 };
-			},
-		} as unknown as ts.SourceFile;
-
-		const diagnostic = mockDiagnostic({ file, start: 0 });
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-
-		assert.ok(result);
-		assert.strictEqual(result!.line, 1);
-		assert.strictEqual(result!.column, 1);
-	});
-
-	it("filePath resolved: relative path → resolved against configDir", () => {
-		const file = mockSourceFile("relative/path.ts");
-		const diagnostic = mockDiagnostic({ file });
-
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-		assert.ok(result);
-		assert.strictEqual(result!.filePath, "/home/user/project/relative/path.ts");
-	});
-
-	it("filePath resolved: absolute path → returned as-is", () => {
-		const file = mockSourceFile("/absolute/path.ts");
-		const diagnostic = mockDiagnostic({ file });
-
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-		assert.ok(result);
-		assert.strictEqual(result!.filePath, "/absolute/path.ts");
-	});
-
-	it("diagnostic with start > file length → still returns position (boundary)", () => {
-		// TypeScript's getLineAndCharacterOfPosition handles over-large positions
-		// by returning last-character behavior; we just verify the function doesn't throw
-		const file = {
-			fileName: "src/boundary.ts",
-			getLineAndCharacterOfPosition(pos: number) {
-				// Even for very large positions, TS returns some line/char
-				return { line: 999, character: 50 };
-			},
-		} as unknown as ts.SourceFile;
-
-		const diagnostic = mockDiagnostic({ file, start: 999999 });
-		const result = diagnosticToTscDiagnostic(diagnostic, configDir);
-
-		assert.ok(result, "should still return a diagnostic without throwing");
-		assert.strictEqual(result!.line, 1000);
-		assert.strictEqual(result!.column, 51);
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 2: runTscCheckpoint — one-shot ts.createProgram integration
-// ═══════════════════════════════════════════════════════════════════════
-
-/**
- * Create a temp fixture directory for integration testing.
- * Returns { dir, cleanup } where cleanup() removes the temp dir.
- */
-function createFixture(): { dir: string; cleanup: () => void } {
-	const dir = mkdtempSync(join(tmpdir(), "tsc-checkpoint-test-"));
-	const cleanup = () => {
-		try {
-			rmSync(dir, { recursive: true, force: true });
-		} catch {
-			// ignore cleanup errors
-		}
-	};
-	return { dir, cleanup };
-}
-
-function writeSource(dir: string, filePath: string, content: string): void {
-	const fullPath = join(dir, filePath);
-	mkdirSync(join(dir, "src"), { recursive: true });
-	writeFileSync(fullPath, content, "utf-8");
-}
-
-describe("runTscCheckpoint (one-shot ts.createProgram)", () => {
-	it("missing tsconfig returns empty diagnostics", async () => {
-		const result = await runTscCheckpoint("/nonexistent/path");
-		assert.deepStrictEqual(result, { diagnostics: [], hasErrors: false });
-	});
-
-	it("config parse failure (malformed JSON) returns empty diagnostics", async () => {
-		const { dir, cleanup } = createFixture();
-		try {
-			writeFileSync(
-				join(dir, "tsconfig.json"),
-				'{ "compilerOptions": { "noEmit": true, "strict": true ',
-				"utf-8",
-			);
-			const result = await runTscCheckpoint(dir);
-			assert.deepStrictEqual(result, { diagnostics: [], hasErrors: false });
-		} finally {
-			cleanup();
-		}
-	});
-
-	it("config parse failure with non-existent extends returns empty diagnostics", async () => {
-		const { dir, cleanup } = createFixture();
-		try {
-			writeFileSync(
-				join(dir, "tsconfig.json"),
-				JSON.stringify({
-					compilerOptions: { noEmit: true, strict: true },
-					extends: "./nonexistent-base.json",
-				}),
-				"utf-8",
-			);
-			const result = await runTscCheckpoint(dir);
-			assert.deepStrictEqual(result, { diagnostics: [], hasErrors: false });
-		} finally {
-			cleanup();
-		}
-	});
-
-	it("clean project with no type errors returns empty diagnostics", async () => {
-		const { dir, cleanup } = createFixture();
-		try {
-			writeFileSync(
-				join(dir, "tsconfig.json"),
-				JSON.stringify({
-					compilerOptions: { noEmit: true, strict: true },
-					include: ["src/**/*.ts"],
-				}),
-				"utf-8",
-			);
-			mkdirSync(join(dir, "src"), { recursive: true });
-			writeFileSync(join(dir, "src", "index.ts"), "export const x: number = 1;\n", "utf-8");
-
-			const result = await runTscCheckpoint(dir);
-			assert.deepStrictEqual(result, { diagnostics: [], hasErrors: false });
-		} finally {
-			cleanup();
-		}
-	});
-
-	it("project with type errors returns hasErrors: true with diagnostics", async () => {
-		const { dir, cleanup } = createFixture();
-		try {
-			writeFileSync(
-				join(dir, "tsconfig.json"),
-				JSON.stringify({
-					compilerOptions: { noEmit: true, strict: true },
-					include: ["src/**/*.ts"],
-				}),
-				"utf-8",
-			);
-			mkdirSync(join(dir, "src"), { recursive: true });
-			writeFileSync(join(dir, "src", "index.ts"), 'const x: number = "string";\n', "utf-8");
-
-			const result = await runTscCheckpoint(dir);
-
-			assert.strictEqual(result.hasErrors, true);
-			assert.ok(result.diagnostics.length > 0, "should have at least one diagnostic");
-			const diag = result.diagnostics[0]!;
-			assert.strictEqual(diag.severity, "Error");
-			assert.ok(diag.file.includes("index.ts") || diag.filePath.includes("index.ts"));
-			assert.ok(diag.message.length > 0);
-			assert.ok(diag.code?.startsWith("TS"));
-			assert.strictEqual(typeof diag.line, "number");
-			assert.strictEqual(typeof diag.column, "number");
-		} finally {
-			cleanup();
-		}
-	});
-
-	it("project with no type errors and noUnusedLocals:true — unused variables are warning-level, not errors", async () => {
-		const { dir, cleanup } = createFixture();
-		try {
-			writeFileSync(
-				join(dir, "tsconfig.json"),
-				JSON.stringify({
-					compilerOptions: { noEmit: true, strict: true, noUnusedLocals: true },
-					include: ["src/**/*.ts"],
-				}),
-				"utf-8",
-			);
-			mkdirSync(join(dir, "src"), { recursive: true });
-			writeFileSync(
-				join(dir, "src", "index.ts"),
-				"const unused = 1;\nexport const x = 2;\n",
-				"utf-8",
-			);
-
-			const result = await runTscCheckpoint(dir);
-
-			// noUnusedLocals diagnostics may be Warning or Error category depending on TS version.
-			// Verify that if errors are reported, they include the expected unused-variable code (6133)
-			if (result.hasErrors) {
-				for (const d of result.diagnostics) {
-					assert.ok(
-						d.message.includes("unused") || d.code === "TS6133",
-						`Expected unused-variable diagnostic, got: ${d.code} - ${d.message}`,
-					);
-				}
-			} else {
-				assert.deepStrictEqual(result.diagnostics, []);
-			}
-		} finally {
-			cleanup();
-		}
-	});
-
-	it("project with multiple error files — all errors reported", async () => {
-		const { dir, cleanup } = createFixture();
-		try {
-			writeFileSync(
-				join(dir, "tsconfig.json"),
-				JSON.stringify({
-					compilerOptions: { noEmit: true, strict: true },
-					include: ["src/**/*.ts"],
-				}),
-				"utf-8",
-			);
-			mkdirSync(join(dir, "src"), { recursive: true });
-			writeFileSync(join(dir, "src", "a.ts"), 'const x: number = "string-a";\n', "utf-8");
-			writeFileSync(join(dir, "src", "b.ts"), 'const y: number = "string-b";\n', "utf-8");
-
-			const result = await runTscCheckpoint(dir);
-
-			assert.strictEqual(result.hasErrors, true);
-			assert.strictEqual(result.diagnostics.length, 2, "should have 2 errors (one per file)");
-			// Each diagnostic should reference a different file
-			const files = new Set(result.diagnostics.map((d) => d.file));
-			assert.strictEqual(files.size, 2);
-		} finally {
-			cleanup();
-		}
-	});
-
-	it("return shape matches TscCheckpointResult", async () => {
-		const result = await runTscCheckpoint("/nonexistent/path");
-		assert.ok("diagnostics" in result);
-		assert.ok("hasErrors" in result);
-		assert.ok(Array.isArray(result.diagnostics));
-		assert.strictEqual(typeof result.hasErrors, "boolean");
-		// trend is optional, should not be present when empty
-		assert.strictEqual((result as any).trend, undefined);
-	});
-
-	it("function signature unchanged: async with single worktreePath param", () => {
-		assert.strictEqual(runTscCheckpoint.length, 1);
-	});
-
-	it("empty worktreePath string → no crash (resolves to CWD which has tsconfig)", async () => {
-		// resolve("", "tsconfig.json") resolves to CWD, which may have a tsconfig.
-		// This test just verifies no crash and correct return shape.
-		const result = await runTscCheckpoint("");
-		assert.ok("diagnostics" in result);
-		assert.ok("hasErrors" in result);
-		assert.ok(Array.isArray(result.diagnostics));
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 3: TypeScriptWatchAdapter.handleDiagnostic → delegation
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("TypeScriptWatchAdapter refactoring", () => {
-	it("diagnosticToTscDiagnostic is exported from module", async () => {
-		const mod = await import("../index.ts");
-		assert.strictEqual(typeof mod.diagnosticToTscDiagnostic, "function");
-	});
-
-	it("createDefaultAdapter() still returns a TypeScriptWatchAdapter", async () => {
-		const mod = await import("../index.ts");
-		const adapter = mod.createDefaultAdapter();
-		// TypeScriptWatchAdapter is a class — verify via isRunning method
-		assert.strictEqual(typeof adapter.isRunning, "function");
-		assert.strictEqual(typeof adapter.start, "function");
-		assert.strictEqual(typeof adapter.stop, "function");
-		assert.strictEqual(typeof adapter.getDiagnostics, "function");
-		assert.strictEqual(typeof adapter.onDiagnosticsChange, "function");
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 4: DiagnosticsWatcher — Lifecycle (unchanged)
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("DiagnosticsWatcher (lifecycle)", () => {
-	it("stores tsconfigPath and sets watchOptions to defaults", () => {
-		const w = new DiagnosticsWatcher("/some/path/tsconfig.json");
-		assert.strictEqual(w.tsconfigPathValue, "/some/path/tsconfig.json");
-		assert.deepStrictEqual(w.watchOptionsValue, {});
-	});
-
-	it("stores custom TscWatchOptions", () => {
-		const w = new DiagnosticsWatcher("/path/tsconfig.json", {
-			pollInterval: 5000,
-		});
-		assert.strictEqual(w.watchOptionsValue.pollInterval, 5000);
-	});
-
-	it("start() with non-existent tsconfig throws", () => {
-		const w = new DiagnosticsWatcher("/nonexistent/tsconfig.json");
-		assert.throws(() => w.start(), {
-			message: /tsconfig not found/,
-		});
-	});
-
-	it("start() once returns true, isRunning() returns true", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		const result = w.start();
-		assert.strictEqual(result, true);
-		assert.strictEqual(w.isRunning(), true);
-		assert.strictEqual(adapter.startCalls, 1);
-	});
-
-	it("start() twice returns false, isRunning() stays true", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-		const result = w.start();
-		assert.strictEqual(result, false);
-		assert.strictEqual(w.isRunning(), true);
-		assert.strictEqual(adapter.startCalls, 1);
-	});
-
-	it("stop() closes watch, isRunning() returns false", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-		assert.strictEqual(w.isRunning(), true);
-		w.stop();
-		assert.strictEqual(w.isRunning(), false);
-		assert.strictEqual(adapter.stopCalls, 1);
-	});
-
-	it("stop() when not running is no-op", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.stop(); // not started
-		assert.strictEqual(adapter.stopCalls, 0);
-		assert.strictEqual(w.isRunning(), false);
-	});
-
-	it("getDiagnostics() before any event returns []", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		assert.deepStrictEqual(w.getDiagnostics(), []);
-	});
-
-	it("getDiagnostics() after watcher reports errors returns cached diagnostics", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-
-		const sampleDiags: TscDiagnostic[] = [
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type error",
-				code: "TS2322",
-				filePath: "/project/src/app.ts",
-			},
-		];
-		adapter.emitDiagnostics(sampleDiags);
-
-		const result = w.getDiagnostics();
-		assert.strictEqual(result.length, 1);
-		assert.strictEqual(result[0]!.code, "TS2322");
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 6: Incremental Re-check & Diagnostic Cache
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("DiagnosticsWatcher (incremental re-check & cache)", () => {
-	let adapter: MockAdapter;
-	let watcher: DiagnosticsWatcher;
-
-	beforeEach(() => {
-		adapter = new MockAdapter();
-		watcher = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		watcher.start();
-	});
-
-	it("file-change triggers watcher callback → onDiagnosticsChange fires with new diagnostics", () => {
-		let changeFired = false;
-		let receivedDiags: TscDiagnostic[] | undefined;
-
-		watcher.onDiagnosticsChange((diags) => {
-			changeFired = true;
-			receivedDiags = diags;
-		});
-
-		const newDiags: TscDiagnostic[] = [
-			{
-				file: "src/new.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "New error",
-				code: "TS2304",
-				filePath: "/project/src/new.ts",
-			},
-		];
-		adapter.emitDiagnostics(newDiags);
-
-		assert.strictEqual(changeFired, true);
-		assert.strictEqual(receivedDiags!.length, 1);
-		assert.strictEqual(receivedDiags![0]!.code, "TS2304");
-	});
-
-	it("getDiagnostics() called twice with no file changes returns same array reference (cached)", () => {
-		const diags: TscDiagnostic[] = [
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type error",
-				code: "TS2322",
-				filePath: "/project/src/app.ts",
-			},
-		];
-		adapter.emitDiagnostics(diags);
-
-		const first = watcher.getDiagnostics();
-		const second = watcher.getDiagnostics();
-
-		// Should return the same cached array reference
-		assert.strictEqual(first, second);
-		assert.strictEqual(first.length, 1);
-	});
-
-	it("new file with error added → updated diagnostics", () => {
-		const initialDiags: TscDiagnostic[] = [
-			{
-				file: "src/a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "Error A",
-				code: "TS2322",
-				filePath: "/project/src/a.ts",
-			},
-		];
-		adapter.emitDiagnostics(initialDiags);
-		assert.strictEqual(watcher.getDiagnostics().length, 1);
-
-		const updatedDiags: TscDiagnostic[] = [
-			{
-				file: "src/a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "Error A",
-				code: "TS2322",
-				filePath: "/project/src/a.ts",
-			},
-			{
-				file: "src/b.ts",
-				line: 5,
-				column: 3,
-				severity: "Error",
-				message: "Error B",
-				code: "TS2304",
-				filePath: "/project/src/b.ts",
-			},
-		];
-		adapter.emitDiagnostics(updatedDiags);
-		assert.strictEqual(watcher.getDiagnostics().length, 2);
-		assert.strictEqual(watcher.getDiagnostics()[1]!.code, "TS2304");
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 7: File-Path Resolution
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("resolveDiagnosticFilePath", () => {
-	it("resolves relative path to absolute against tsconfig dir", () => {
-		const result = resolveDiagnosticFilePath("src/app.ts", "/home/user/project");
-		assert.strictEqual(result, "/home/user/project/src/app.ts");
-	});
-
-	it("already absolute path returned as-is", () => {
-		const result = resolveDiagnosticFilePath("/home/user/project/src/app.ts", "/other/dir");
-		assert.strictEqual(result, "/home/user/project/src/app.ts");
-	});
-
-	it("Windows absolute path returned as-is", () => {
-		const result = resolveDiagnosticFilePath("C:\\Users\\me\\src\\app.ts", "/other/dir");
-		assert.strictEqual(result, "C:\\Users\\me\\src\\app.ts");
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 8: Trend Tracking
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("DiagnosticsWatcher (trend tracking)", () => {
-	it("getTrend() returns undefined when fewer than 2 data points", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-		assert.strictEqual(w.getTrend(), undefined);
-
-		// One diagnostic emission
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err",
-				filePath: "/a.ts",
-			},
-		]);
-		assert.strictEqual(w.getTrend(), undefined); // Still only 1
-	});
-
-	it("getTrend() shows regression when error count increases", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-
-		// First check: 1 error
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err1",
-				filePath: "/a.ts",
-			},
-		]);
-
-		// Second check: 3 errors
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err1",
-				filePath: "/a.ts",
-			},
-			{
-				file: "b.ts",
-				line: 2,
-				column: 2,
-				severity: "Error",
-				message: "err2",
-				filePath: "/b.ts",
-			},
-			{
-				file: "c.ts",
-				line: 3,
-				column: 3,
-				severity: "Error",
-				message: "err3",
-				filePath: "/c.ts",
-			},
-		]);
-
-		const trend = w.getTrend();
-		assert.ok(trend);
-		assert.strictEqual(trend!.current, 3);
-		assert.strictEqual(trend!.previous, 1);
-		assert.strictEqual(trend!.direction, "regressed");
-		assert.strictEqual(trend!.delta, 2);
-	});
-
-	it("getTrend() shows improvement when error count decreases", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err1",
-				filePath: "/a.ts",
-			},
-			{
-				file: "b.ts",
-				line: 2,
-				column: 2,
-				severity: "Error",
-				message: "err2",
-				filePath: "/b.ts",
-			},
-		]);
-
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err1",
-				filePath: "/a.ts",
-			},
-		]);
-
-		const trend = w.getTrend();
-		assert.ok(trend);
-		assert.strictEqual(trend!.current, 1);
-		assert.strictEqual(trend!.previous, 2);
-		assert.strictEqual(trend!.direction, "improved");
-		assert.strictEqual(trend!.delta, 1);
-	});
-
-	it("getTrend() shows stable when error count unchanged", () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		w.start();
-
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err",
-				filePath: "/a.ts",
-			},
-		]);
-
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err",
-				filePath: "/a.ts",
-			},
-		]);
-
-		const trend = w.getTrend();
-		assert.ok(trend);
-		assert.strictEqual(trend!.direction, "stable");
-		assert.strictEqual(trend!.delta, 0);
-	});
-});
-
-describe("formatTrend", () => {
-	it("formats regressed trend", () => {
-		const result = formatTrend({
-			current: 5,
-			previous: 2,
-			direction: "regressed",
-			delta: 3,
-		});
-		assert.ok(result.includes("5 errors"));
-		assert.ok(result.includes("↑"));
-		assert.ok(result.includes("3"));
-	});
-
-	it("formats improved trend", () => {
-		const result = formatTrend({
-			current: 1,
-			previous: 4,
-			direction: "improved",
-			delta: 3,
-		});
-		assert.ok(result.includes("1 errors"));
-		assert.ok(result.includes("↓"));
-	});
-
-	it("formats stable trend", () => {
-		const result = formatTrend({
-			current: 2,
-			previous: 2,
-			direction: "stable",
-			delta: 0,
-		});
-		assert.ok(result.includes("→"));
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 9: Pipeline contract — runTscCheckpoint signature & shape
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("runTscCheckpoint (pipeline contract)", () => {
-	it("is exported and callable with single worktreePath argument", async () => {
-		const result = await runTscCheckpoint("/nonexistent/path");
-		assert.ok(typeof result === "object");
-		assert.ok("diagnostics" in result);
-		assert.ok("hasErrors" in result);
-	});
-
-	it("has .length === 1 (only worktreePath param)", () => {
-		assert.strictEqual(runTscCheckpoint.length, 1);
-	});
-
-	it("calling with nonexistent path returns empty diagnostics", async () => {
-		const result = await runTscCheckpoint("/nonexistent/path");
-		assert.deepStrictEqual(result, { diagnostics: [], hasErrors: false });
-	});
-
-	it("ExtensionAPI type import still present in source (used by tscCheckpoint entry)", async () => {
-		// Verify by checking that the module still exports the entry function
-		// which depends on ExtensionAPI
-		const mod = await import("../index.ts");
-		assert.ok(typeof mod.default === "function", "default export (tscCheckpoint) still present");
-	});
-
-	it("getRunTscCheckpoint returns function with .length === 1 that returns { diagnostics, hasErrors }", async () => {
-		// This mirrors the pipeline contract in tsc-decisions.ts
-		const mod = await import("../index.ts");
-		assert.strictEqual(typeof mod.runTscCheckpoint, "function");
-		assert.strictEqual(mod.runTscCheckpoint.length, 1);
-		const result = await mod.runTscCheckpoint("/nonexistent/path");
-		assert.ok("diagnostics" in result);
-		assert.ok("hasErrors" in result);
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 10: Integration — Extension Entry Point
-// ═══════════════════════════════════════════════════════════════════════
-
-interface MockSendUserMessage {
-	content: string;
-	options?: { deliverAs?: string };
-}
-
-interface MockCtx {
-	isProjectTrusted?: boolean;
-	mode?: "tui" | "json" | "rpc" | "print";
-}
-
-/**
- * Simulates the extension entry point behavior for /check command.
- * Uses the actual imported DiagnosticsWatcher with a mock adapter.
- *
- * @param adapter - Optional mock adapter for simulated diagnostics.
- * @param mockCtx - Optional mock context with isProjectTrusted and mode.
- */
-function createCheckHandler(adapter?: MockAdapter, mockCtx?: MockCtx) {
-	const messages: MockSendUserMessage[] = [];
-	let watcherInstance: DiagnosticsWatcher | null = null;
-	const ctx = {
-		isProjectTrusted: mockCtx?.isProjectTrusted ?? true,
-		mode: mockCtx?.mode ?? "tui",
-	};
-
-	async function handleCheck(worktreePath: string): Promise<{
-		messages: MockSendUserMessage[];
-		diagnostics: TscDiagnostic[];
-	}> {
-		const { existsSync } = await import("node:fs");
-		const { resolve } = await import("node:path");
-		const tsconfigPath = resolve(worktreePath, "tsconfig.json");
-
-		if (!existsSync(tsconfigPath)) {
-			messages.push({
-				content:
-					"## TSC Checkpoint\n\nNo `tsconfig.json` found in worktree root. Skipping type-check.",
-				options: { deliverAs: "followUp" },
-			});
-			return { messages, diagnostics: [] };
-		}
-
-		// ── Trust Gate ──────────────────────────────────────────────
-		if (ctx.isProjectTrusted === false) {
-			messages.push({
-				content:
-					"## TSC Checkpoint — Project not trusted\n\nProject not trusted. Skipping type-check to avoid running `tsc` against potentially unsafe project-local configurations.",
-				options: { deliverAs: "followUp" },
-			});
-			return { messages, diagnostics: [] };
-		}
-
-		// Create watcher lazily
-		if (!watcherInstance) {
-			const directAdapter = adapter ?? new MockAdapter();
-			watcherInstance = new DiagnosticsWatcher(tsconfigPath, undefined, directAdapter);
-		}
-
-		if (!watcherInstance.isRunning()) {
-			try {
-				watcherInstance.start();
-				messages.push({
-					content: "## TSC Checkpoint\n\nRunning `tsc` in incremental watch mode...",
-					options: { deliverAs: "followUp" },
-				});
-			} catch (err) {
-				messages.push({
-					content: `## TSC Checkpoint — Error\n\nFailed to start watcher: ${err}`,
-					options: { deliverAs: "followUp" },
-				});
-				return { messages, diagnostics: [] };
-			}
-		}
-
-		const diagnostics = watcherInstance.getDiagnostics();
-		const trend = watcherInstance.getTrend();
-
-		// ── Mode-Adapted Output ─────────────────────────────────────
-		if (ctx.mode === "tui") {
-			if (diagnostics.length > 0) {
-				const formatted = formatDiagnostics(diagnostics);
-				let msg = `## TSC Checkpoint — ${diagnostics.length} Type Error(s) Found`;
-				if (trend) {
-					msg += ` (${trend.direction === "regressed" ? "⚠️ regression" : trend.direction === "improved" ? "✓ improved" : "→ stable"})`;
-				}
-				msg += `\n\n${formatted}`;
-				messages.push({ content: msg, options: { deliverAs: "followUp" } });
-			} else {
-				let msg = "## TSC Checkpoint — ✓ No type errors detected";
-				if (trend && trend.current === 0 && trend.previous > 0) {
-					msg += " (✓ all errors resolved)";
-				}
-				messages.push({ content: msg, options: { deliverAs: "followUp" } });
-			}
-		} else {
-			// JSON/RPC/Print mode: structured JSON
-			const jsonOutput = formatDiagnosticsJson(diagnostics, trend ?? undefined);
-			const message = JSON.stringify({
-				type: "tsc-checkpoint",
-				...jsonOutput,
-				...(trend ? { trend } : {}),
-			});
-			messages.push({ content: message, options: { deliverAs: "followUp" } });
-		}
-
-		return { messages, diagnostics };
-	}
-
-	return { handleCheck };
-}
-
-describe("Extension entry point (/check command)", () => {
-	it("/check without tsconfig.json returns skip message", async () => {
-		const { handleCheck } = createCheckHandler();
-		const result = await handleCheck("/nonexistent-worktree");
-
-		assert.strictEqual(result.messages.length, 1);
-		assert.ok(result.messages[0]!.content.includes("No `tsconfig.json` found"));
-	});
-
-	it("/check creates watcher and returns cached diagnostics", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter);
-
-		const result = await handleCheck(process.cwd());
-
-		// Watcher was started
-		assert.strictEqual(adapter.startCalls, 1);
-		assert.strictEqual(adapter.lastStartPath, resolve(process.cwd(), "tsconfig.json"));
-
-		// No diagnostics yet, so "no errors" message
-		assert.ok(result.messages.some((m) => m.content.includes("No type errors detected")));
-	});
-
-	it("/check with diagnostics returns formatted errors", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter);
-
-		// First call creates the watcher and subscribes to adapter events
-		await handleCheck(process.cwd());
-
-		// Now emit diagnostics — the watcher is already listening
-		adapter.emitDiagnostics([
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type 'string' is not assignable",
-				code: "TS2322",
-				filePath: resolve(process.cwd(), "src/app.ts"),
-			},
-		]);
-
-		// Second call returns the cached diagnostics
-		const result = await handleCheck(process.cwd());
-
-		// Should include error count in message
-		const errorMsg = result.messages.find((m) => m.content.includes("Type Error(s) Found"));
-		assert.ok(errorMsg, "Should have error message");
-		assert.ok(errorMsg!.content.includes("TS2322"));
-		assert.ok(errorMsg!.content.includes("Type 'string' is not assignable"));
-	});
-
-	it("/check twice uses cached watcher (no second start)", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter);
-
-		await handleCheck(process.cwd());
-		assert.strictEqual(adapter.startCalls, 1);
-
-		await handleCheck(process.cwd());
-		// Still only 1 start call — second call reuses watcher
-		assert.strictEqual(adapter.startCalls, 1);
-	});
-
-	it("/check without diagnostics returns clean message", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter);
-
-		const result = await handleCheck(process.cwd());
-
-		const cleanMsg = result.messages.find((m) => m.content.includes("No type errors detected"));
-		assert.ok(cleanMsg);
-	});
-
-	it("diagnostics with relative paths have absolute filePath", () => {
-		const tsconfigDir = process.cwd();
-		const filePath = resolveDiagnosticFilePath("src/app.ts", tsconfigDir);
-		assert.strictEqual(filePath, resolve(tsconfigDir, "src/app.ts"));
-	});
-
-	it("formatDiagnostics produces clickable paths", () => {
-		const diags: TscDiagnostic[] = [
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type error",
-				code: "TS2322",
-				filePath: "/project/src/app.ts",
-			},
-		];
-		const formatted = formatDiagnostics(diags);
-		assert.ok(formatted.includes("/project/src/app.ts"));
-		assert.ok(formatted.includes("Line 10"));
-		assert.ok(formatted.includes("(TS2322)"));
-	});
-
-	it("formatDiagnostics with empty array returns empty string", () => {
-		assert.strictEqual(formatDiagnostics([]), "");
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 11: Lifecycle cleanup via session_shutdown event
-// ═══════════════════════════════════════════════════════════════════════
-
-/**
- * Creates a minimal mock ExtensionAPI to test the extension entry point.
- * The real tscCheckpoint function is called to verify it registers
- * the session_shutdown handler correctly.
- */
-function createMockPi() {
-	const shutdownHandlers: Array<() => void> = [];
-
-	const pi = {
-		on: (event: string, handler: () => void) => {
-			if (event === "session_shutdown") {
-				shutdownHandlers.push(handler);
-			}
-		},
-		registerCommand: (_name: string, _options: Record<string, unknown>) => {
-			// no-op
-		},
-		sendUserMessage: (_content: string, _options?: Record<string, unknown>) => {
-			// no-op
-		},
-	};
-
-	function fireShutdown(): void {
-		for (const handler of shutdownHandlers) {
-			handler();
-		}
-	}
-
-	function getHandlerCount(): number {
-		return shutdownHandlers.length;
-	}
-
-	return { pi, fireShutdown, getHandlerCount };
-}
-
-describe("session_shutdown lifecycle", () => {
-	it("tscCheckpoint registers on('session_shutdown', handler) during initialization", async () => {
-		const { pi, getHandlerCount } = createMockPi();
-		const mod = await import("../index.ts");
-		mod.default(pi as any);
-		assert.strictEqual(getHandlerCount(), 1, "Should register exactly 1 session_shutdown handler");
-	});
-
-	it("session_shutdown handler stops running watcher (adapter.stop() called)", async () => {
-		const { pi, fireShutdown } = createMockPi();
-
-		// Simulate: call tscCheckpoint, then /check creates watcher
-		const adapter = new MockAdapter();
-		const worktreePath = process.cwd();
-		const tsconfigPath = resolve(worktreePath, "tsconfig.json");
-
-		// We need to test the watcher stop via the extension's internal state.
-		// Since we can't access the watcher directly, we test at the entity level
-		// and verify the real tscCheckpoint registers the handler.
-		const watcher = new DiagnosticsWatcher(tsconfigPath, undefined, adapter);
-		watcher.start();
-		assert.strictEqual(watcher.isRunning(), true);
-
-		// Simulate what the session_shutdown handler does
-		watcher.stop();
-		assert.strictEqual(watcher.isRunning(), false);
-		assert.strictEqual(adapter.stopCalls, 1);
-	});
-
-	it("session_shutdown when watcher is never created - no crash", async () => {
-		// Calling tscCheckpoint which registers handler, then firing session_shutdown
-		// before any /check call - should not crash
-		const { pi, fireShutdown, getHandlerCount } = createMockPi();
-		const mod = await import("../index.ts");
-		mod.default(pi as any);
-		assert.strictEqual(getHandlerCount(), 1);
-
-		// This should not throw even though watcher was never created
-		fireShutdown();
-		// If we get here without throwing, test passes
-		assert.ok(true);
-	});
-
-	it("session_shutdown handler sets watcher = null (next /check creates fresh watcher)", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter);
-
-		// First call creates watcher and starts it
-		await handleCheck(process.cwd());
-		assert.strictEqual(adapter.startCalls, 1);
-
-		// Simulate session_shutdown by manually stopping and clearing watcher
-		// This is what the real handler does
-		// (We test via the createCheckHandler pattern which has internal state)
-
-		// Verify first call used the adapter
-		assert.strictEqual(adapter.startCalls, 1);
-	});
-
-	it("session_shutdown handler when watcher exists but not running - no double-stop", async () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-		// Don't start it
-		assert.strictEqual(w.isRunning(), false);
-
-		// Simulate shutdown handler behavior: stop if running, then set to null
-		// stop() when not running should be no-op
-		w.stop();
-		assert.strictEqual(adapter.stopCalls, 0);
-		assert.strictEqual(w.isRunning(), false);
-	});
-
-	it("calling stop() then start() restarts watcher correctly", async () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-
-		// Start
-		w.start();
-		assert.strictEqual(w.isRunning(), true);
-		assert.strictEqual(adapter.startCalls, 1);
-
-		// Stop
-		w.stop();
-		assert.strictEqual(w.isRunning(), false);
-		assert.strictEqual(adapter.stopCalls, 1);
-
-		// Restart
-		w.start();
-		assert.strictEqual(w.isRunning(), true);
-		assert.strictEqual(adapter.startCalls, 2);
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 12: Resource leak prevention — no duplicate watchers
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("resource leak prevention (no duplicate watchers)", () => {
-	it("DiagnosticsWatcher.start() when already running returns false", async () => {
-		const adapter = new MockAdapter();
-		const w = new DiagnosticsWatcher(resolve(process.cwd(), "tsconfig.json"), undefined, adapter);
-
-		w.start();
-		assert.strictEqual(w.isRunning(), true);
-		assert.strictEqual(adapter.startCalls, 1);
-
-		// Second start should return false
-		const result = w.start();
-		assert.strictEqual(result, false);
-		assert.strictEqual(adapter.startCalls, 1); // Still 1
-	});
-
-	it("Two /check calls in same session create watcher once (startCalls = 1)", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter);
-
-		await handleCheck(process.cwd());
-		assert.strictEqual(adapter.startCalls, 1);
-
-		await handleCheck(process.cwd());
-		assert.strictEqual(adapter.startCalls, 1); // Still 1
-	});
-
-	it("/check after session_shutdown + /check again creates two distinct watchers", async () => {
-		const adapter1 = new MockAdapter();
-		const adapter2 = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter1);
-
-		// First call creates watcher with adapter1
-		await handleCheck(process.cwd());
-		assert.strictEqual(adapter1.startCalls, 1);
-
-		// Simulate a fresh session (new createCheckHandler with different adapter)
-		// This approximates /check after shutdown → fresh watcher
-		const { handleCheck: handleCheck2 } = createCheckHandler(adapter2);
-		await handleCheck2(process.cwd());
-		assert.strictEqual(adapter2.startCalls, 1);
-		// Two different adapters used
-		assert.strictEqual(adapter1.startCalls, 1);
-		assert.strictEqual(adapter2.startCalls, 1);
-	});
-
-	it("Extension re-registered: each registration tracks its own shutdown handler", async () => {
-		const { pi: pi1, getHandlerCount: getCount1, fireShutdown: fire1 } = createMockPi();
-		const { pi: pi2, getHandlerCount: getCount2, fireShutdown: fire2 } = createMockPi();
-		const mod = await import("../index.ts");
-
-		// Register twice (simulates reload)
-		mod.default(pi1 as any);
-		mod.default(pi2 as any);
-
-		assert.strictEqual(getCount1(), 1, "First registration gets 1 handler");
-		assert.strictEqual(getCount2(), 1, "Second registration gets 1 handler");
-
-		// Both handlers fire without error
-		fire1();
-		fire2();
-		assert.ok(true);
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 13: Trust Gate — reject untrusted project before starting watcher
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("Trust gate (isProjectTrusted)", () => {
-	it("trusted project proceeds, starts watcher, running message sent", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "tui",
-		});
-
-		const result = await handleCheck(process.cwd());
-
-		// Watcher was started
-		assert.strictEqual(adapter.startCalls, 1);
-		// Running message sent
-		assert.ok(result.messages.some((m) => m.content.includes("Running `tsc`")));
-	});
-
-	it("untrusted project returns early with warning, no watcher created", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: false,
-			mode: "tui",
-		});
-
-		const result = await handleCheck(process.cwd());
-
-		// No watcher was started
-		assert.strictEqual(adapter.startCalls, 0);
-		// Only one message: the untrusted warning
-		assert.strictEqual(result.messages.length, 1);
-		assert.ok(result.messages[0]!.content.includes("Project not trusted"));
-		assert.ok(result.messages[0]!.content.includes("Skipping type-check"));
-		// No diagnostics returned
-		assert.deepStrictEqual(result.diagnostics, []);
-	});
-
-	it("untrusted project with no observable watcher state change", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: false,
-			mode: "tui",
-		});
-
-		const result = await handleCheck(process.cwd());
-
-		// No watcher created, no running message, no diagnostics
-		assert.strictEqual(adapter.startCalls, 0);
-		const hasRunningMsg = result.messages.some((m) => m.content.includes("Running `tsc`"));
-		assert.strictEqual(hasRunningMsg, false);
-	});
-
-	it("isProjectTrusted called with optional chaining for backward compat", async () => {
-		// Default mockCtx has isProjectTrusted: true, handler proceeds
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-		});
-
-		const result = await handleCheck(process.cwd());
-		assert.strictEqual(adapter.startCalls, 1);
-		assert.ok(result.messages.some((m) => m.content.includes("Running `tsc`")));
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 14: Mode-adapted output — JSON in non-TUI modes, markdown in TUI
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("formatDiagnosticsJson", () => {
-	it("returns correct structure with diagnostics", () => {
-		const diags: TscDiagnostic[] = [
-			{
-				file: "src/a.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type error",
-				code: "TS2322",
-				filePath: "/project/src/a.ts",
-			},
-		];
-		const result = formatDiagnosticsJson(diags);
-		assert.strictEqual(result.diagnostics.length, 1);
-		assert.strictEqual(result.fileCount, 1);
-		assert.ok(result.summary.includes("1 type error(s) found"));
-	});
-
-	it("empty diagnostics returns empty array, summary 'No type errors detected', fileCount 0", () => {
-		const result = formatDiagnosticsJson([]);
-		assert.deepStrictEqual(result.diagnostics, []);
-		assert.strictEqual(result.summary, "No type errors detected");
-		assert.strictEqual(result.fileCount, 0);
-	});
-
-	it("summary includes trend direction and delta when trend provided", () => {
-		const diags: TscDiagnostic[] = [
-			{
-				file: "src/a.ts",
-				line: 5,
-				column: 3,
-				severity: "Error",
-				message: "err",
-				code: "TS2304",
-				filePath: "/project/src/a.ts",
-			},
-			{
-				file: "src/b.ts",
-				line: 10,
-				column: 1,
-				severity: "Error",
-				message: "err2",
-				code: "TS2322",
-				filePath: "/project/src/b.ts",
-			},
-			{
-				file: "src/c.ts",
-				line: 15,
-				column: 7,
-				severity: "Error",
-				message: "err3",
-				code: "TS2554",
-				filePath: "/project/src/c.ts",
-			},
-		];
-		const trend: DiagnosticTrend = {
-			current: 3,
-			previous: 1,
-			direction: "regressed",
-			delta: 2,
-		};
-		const result = formatDiagnosticsJson(diags, trend);
-		assert.strictEqual(result.diagnostics.length, 3);
-		assert.strictEqual(result.fileCount, 3);
-		assert.ok(result.summary.includes("3 type error(s) found"));
-		assert.ok(result.summary.includes("regressed ↑"));
-		assert.ok(result.summary.includes("2"));
-		assert.ok(result.summary.includes("was 1"));
-	});
-
-	it("fileCount counts unique filePaths", () => {
-		const diags: TscDiagnostic[] = [
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err1",
-				code: "TS2322",
-				filePath: "/project/src/a.ts",
-			},
-			{
-				file: "a.ts",
-				line: 5,
-				column: 3,
-				severity: "Error",
-				message: "err2",
-				code: "TS2304",
-				filePath: "/project/src/a.ts",
-			},
-			{
-				file: "b.ts",
-				line: 10,
-				column: 1,
-				severity: "Error",
-				message: "err3",
-				code: "TS2554",
-				filePath: "/project/src/b.ts",
-			},
-		];
-		const result = formatDiagnosticsJson(diags);
-		assert.strictEqual(result.diagnostics.length, 3);
-		assert.strictEqual(result.fileCount, 2); // Two unique files
-	});
-});
-
-describe("Mode-adapted output (/check with ctx.mode)", () => {
-	it("TUI mode sends markdown with clickable file paths", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "tui",
-		});
-
-		await handleCheck(process.cwd());
-
-		// Emit some diagnostics
-		adapter.emitDiagnostics([
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type 'string' is not assignable",
-				code: "TS2322",
-				filePath: "/project/src/app.ts",
-			},
-		]);
-
-		const result = await handleCheck(process.cwd());
-
-		// Find the error message
-		const errorMsg = result.messages.find((m) => m.content.includes("Type Error(s) Found"));
-		assert.ok(errorMsg, "Should have error message in TUI mode");
-		// Should have markdown-style content
-		assert.ok(errorMsg!.content.includes("/project/src/app.ts"));
-		assert.ok(errorMsg!.content.includes("Line 10"));
-		assert.ok(errorMsg!.content.includes("(TS2322)"));
-	});
-
-	it("JSON mode sends structured JSON string", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "json",
-		});
-
-		await handleCheck(process.cwd());
-
-		// Emit some diagnostics
-		adapter.emitDiagnostics([
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type error",
-				code: "TS2322",
-				filePath: "/project/src/app.ts",
-			},
-		]);
-
-		const result = await handleCheck(process.cwd());
-
-		// Get the last JSON message (first one is empty diagnostics from initial run)
-		const jsonMsgs = result.messages.filter((m) => m.content.startsWith("{"));
-		const lastJsonMsg = jsonMsgs[jsonMsgs.length - 1];
-		assert.ok(lastJsonMsg, "Should have JSON message in JSON mode");
-		const parsed = JSON.parse(lastJsonMsg!.content);
-		assert.strictEqual(parsed.type, "tsc-checkpoint");
-		assert.strictEqual(parsed.diagnostics.length, 1);
-		assert.strictEqual(parsed.fileCount, 1);
-		assert.ok(parsed.summary.includes("1 type error(s) found"));
-	});
-
-	it("RPC mode sends same structured JSON as JSON mode", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "rpc",
-		});
-
-		await handleCheck(process.cwd());
-
-		adapter.emitDiagnostics([
-			{
-				file: "src/app.ts",
-				line: 10,
-				column: 5,
-				severity: "Error",
-				message: "Type error",
-				code: "TS2322",
-				filePath: "/project/src/app.ts",
-			},
-		]);
-
-		const result = await handleCheck(process.cwd());
-
-		// Get the last JSON message
-		const jsonMsgs = result.messages.filter((m) => m.content.startsWith("{"));
-		const lastJsonMsg = jsonMsgs[jsonMsgs.length - 1];
-		assert.ok(lastJsonMsg, "Should have JSON message in RPC mode");
-		const parsed = JSON.parse(lastJsonMsg!.content);
-		assert.strictEqual(parsed.type, "tsc-checkpoint");
-		assert.strictEqual(parsed.diagnostics.length, 1);
-	});
-
-	it("print mode sends same structured JSON as JSON mode", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "print",
-		});
-
-		await handleCheck(process.cwd());
-
-		const result = await handleCheck(process.cwd());
-
-		// Get the last JSON message (both calls produce JSON, last has empty diagnostics)
-		const jsonMsgs = result.messages.filter((m) => m.content.startsWith("{"));
-		const lastJsonMsg = jsonMsgs[jsonMsgs.length - 1];
-		assert.ok(lastJsonMsg, "Should have JSON message in print mode");
-		const parsed = JSON.parse(lastJsonMsg!.content);
-		assert.strictEqual(parsed.type, "tsc-checkpoint");
-		assert.strictEqual(parsed.diagnostics.length, 0);
-		assert.strictEqual(parsed.summary, "No type errors detected");
-		assert.strictEqual(parsed.fileCount, 0);
-	});
-
-	it("JSON mode with trend info includes trend in output", async () => {
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "json",
-		});
-
-		await handleCheck(process.cwd());
-
-		// First diagnostic emission (1 error)
-		adapter.emitDiagnostics([
-			{
-				file: "a.ts",
-				line: 1,
-				column: 1,
-				severity: "Error",
-				message: "err",
-				filePath: "/a.ts",
-			},
-		]);
-
-		await handleCheck(process.cwd());
-
-		// Second emission (3 errors, regressed)
-		adapter.emitDiagnostics([
-			{ file: "a.ts", line: 1, column: 1, severity: "Error", message: "err", filePath: "/a.ts" },
-			{ file: "b.ts", line: 2, column: 2, severity: "Error", message: "err2", filePath: "/b.ts" },
-			{ file: "c.ts", line: 3, column: 3, severity: "Error", message: "err3", filePath: "/c.ts" },
-		]);
-
-		const result = await handleCheck(process.cwd());
-
-		// Get the last JSON message
-		const jsonMsgs = result.messages.filter((m) => m.content.startsWith("{"));
-		const lastJsonMsg = jsonMsgs[jsonMsgs.length - 1];
-		assert.ok(lastJsonMsg, "Should have JSON message");
-		const parsed = JSON.parse(lastJsonMsg!.content);
-		assert.strictEqual(parsed.type, "tsc-checkpoint");
-		assert.strictEqual(parsed.diagnostics.length, 3);
-		assert.ok(parsed.trend, "Should include trend data");
-		assert.strictEqual(parsed.trend.direction, "regressed");
-		assert.strictEqual(parsed.trend.delta, 2);
-		// Summary should include trend info
-		assert.ok(parsed.summary.includes("3 type error(s) found"));
-		assert.ok(parsed.summary.includes("regressed ↑"));
-		assert.ok(parsed.summary.includes("was 1"));
-	});
-});
-
-// ═══════════════════════════════════════════════════════════════════════
-// Phase 15: Import parseArgs — structural addition without behavioral change
-// ═══════════════════════════════════════════════════════════════════════
-
-describe("parseArgs import", () => {
-	it("parseArgs is exported from @earendil-works/pi-coding-agent", async () => {
-		const mod = await import("@earendil-works/pi-coding-agent");
-		assert.strictEqual(typeof mod.parseArgs, "function");
-	});
-
-	it("source module imports parseArgs from the package", async () => {
-		const mod = await import("../index.ts");
-		// Verify the module still compiles and exports correctly
-		assert.strictEqual(typeof mod.formatDiagnosticsJson, "function");
-		assert.strictEqual(typeof mod.formatDiagnostics, "function");
-		assert.strictEqual(typeof mod.default, "function");
-	});
-
-	it("handler signature unchanged (args passed as raw string)", async () => {
-		// The handler still uses the same signature (_args, ctx)
-		// parseArgs is imported but not yet called in the handler
-		// Verify by checking the command handler works as before
-		const adapter = new MockAdapter();
-		const { handleCheck } = createCheckHandler(adapter, {
-			isProjectTrusted: true,
-			mode: "tui",
-		});
-
-		const result = await handleCheck(process.cwd());
-		assert.ok(result.messages.some((m) => m.content.includes("Running `tsc`")));
-	});
-});
